requirements.txt
# Listes des librairies à installer
# pip install -r requirements.txt

# Core

sqlalchemy==2.0.35
psycopg[binary]==3.3.2
python-dotenv==1.0.1
alembic==1.13.1

# Data & Scraping

yfinance==0.2.36
pandas==2.1.4
numpy==1.26.2
requests==2.31.0

# Quant / Risk

scipy==1.11.4
py_vollib==1.0.1
statsmodels==0.14.1

# Export / BI

openpyxl==3.1.2
xlsxwriter==3.1.9

# Dev

pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
